Refactor AnalysisBasedWarnings.cpp
